FX-30: Cross-Container Portability (MyFaces support, Geronimo detection in SystemInfoBean)
git-svn-id: http://svn.flexive.org/repository/flexive/trunk@1031 d35d4baa-7912-0410-b0ae-edacad209e15

+++ src/ui/jsf-components/src/java/com/flexive/faces/beans/SelectBean.java
@@ -319,7 +319,7 @@ public class SelectBean implements Serializable {
         UserTicket ticket = FxJsfUtils.getRequest().getUserTicket();
         List<FxSelectList> selectLists = CacheAdmin.getFilteredEnvironment().getSelectLists();
         ArrayList<SelectItem> result = new ArrayList<SelectItem>(selectLists.size() + 1);
-        result.add(new SelectItem((long) -1, ""));
+        result.add(new SelectItem(-1L, ""));
         for (FxSelectList list : selectLists) {
             result.add(new SelectItem(list.getId(), list.getLabel().getBestTranslation(ticket),
                     list.getDescription().getBestTranslation(ticket)));
